rustsbi: guest instance; u32 function and module id

+++ src/ecall/hsm.rs
@@ -1,13 +1,13 @@
 //! hsm extension
 use super::SbiRet;
 
-const FUNCTION_HSM_HART_START: usize = 0x0;
-const FUNCTION_HSM_HART_STOP: usize = 0x1;
-const FUNCTION_HSM_HART_GET_STATUS: usize = 0x2;
-const FUNCTION_HSM_HART_SUSPEND: usize = 0x3;
+const FUNCTION_HSM_HART_START: u32 = 0x0;
+const FUNCTION_HSM_HART_STOP: u32 = 0x1;
+const FUNCTION_HSM_HART_GET_STATUS: u32 = 0x2;
+const FUNCTION_HSM_HART_SUSPEND: u32 = 0x3;
 
 #[inline]
-pub fn handle_ecall_hsm(function: usize, param0: usize, param1: usize, param2: usize) -> SbiRet {
+pub fn handle_ecall_hsm(function: u32, param0: usize, param1: usize, param2: usize) -> SbiRet {
     match function {
         FUNCTION_HSM_HART_START => hart_start(param0, param1, param2),
         FUNCTION_HSM_HART_STOP => hart_stop(param0),
