+++ src/ecall/rfence.rs
@@ -2,17 +2,17 @@ use super::SbiRet;
 use crate::hart_mask::HartMask;
 use crate::rfence;
 
-const FUNCTION_RFENCE_REMOTE_FENCE_I: usize = 0x0;
-const FUNCTION_RFENCE_REMOTE_SFENCE_VMA: usize = 0x1;
-const FUNCTION_RFENCE_REMOTE_SFENCE_VMA_ASID: usize = 0x2;
-const FUNCTION_RFENCE_REMOTE_HFENCE_GVMA_VMID: usize = 0x3;
-const FUNCTION_RFENCE_REMOTE_HFENCE_GVMA: usize = 0x4;
-const FUNCTION_RFENCE_REMOTE_HFENCE_VVMA_ASID: usize = 0x5;
-const FUNCTION_RFENCE_REMOTE_HFENCE_VVMA: usize = 0x6;
+const FUNCTION_RFENCE_REMOTE_FENCE_I: u32 = 0x0;
+const FUNCTION_RFENCE_REMOTE_SFENCE_VMA: u32 = 0x1;
+const FUNCTION_RFENCE_REMOTE_SFENCE_VMA_ASID: u32 = 0x2;
+const FUNCTION_RFENCE_REMOTE_HFENCE_GVMA_VMID: u32 = 0x3;
+const FUNCTION_RFENCE_REMOTE_HFENCE_GVMA: u32 = 0x4;
+const FUNCTION_RFENCE_REMOTE_HFENCE_VVMA_ASID: u32 = 0x5;
+const FUNCTION_RFENCE_REMOTE_HFENCE_VVMA: u32 = 0x6;
 
 #[inline]
 pub fn handle_ecall_rfence(
-    function: usize,
+    function: u32,
     param0: usize,
     param1: usize,
     param2: usize,
