+++ src/ecall/srst.rs
@@ -1,9 +1,9 @@
 use super::SbiRet;
 
-const FUNCTION_SYSTEM_RESET: usize = 0x0;
+const FUNCTION_SYSTEM_RESET: u32 = 0x0;
 
 #[inline]
-pub fn handle_ecall_srst(function: usize, param0: usize, param1: usize) -> SbiRet {
+pub fn handle_ecall_srst(function: u32, param0: usize, param1: usize) -> SbiRet {
     match function {
         FUNCTION_SYSTEM_RESET => system_reset(param0, param1),
         _ => SbiRet::not_supported(),
